v1.8.4
- Remove middleware folder
- Extend travis testing
