+++ go.mod
@@ -3,11 +3,8 @@ module github.com/gofiber/fiber
 go 1.11
 
 require (
-	github.com/fasthttp/websocket v1.4.2
 	github.com/gofiber/template v1.0.0
-	github.com/google/uuid v1.1.1
 	github.com/gorilla/schema v1.1.0
 	github.com/json-iterator/go v1.1.9
 	github.com/valyala/fasthttp v1.9.0
-	github.com/valyala/fasttemplate v1.1.0
 )
